Change activation and create target

+++ src/main/java/nu/aron/next/NextBuildNumberLifecycleParticipant.java
@@ -1,4 +1,4 @@
-package nu.aron.nextbuildnumber;
+package nu.aron.next;
 
 import io.vavr.collection.List;
 import io.vavr.control.Try;
@@ -13,14 +13,20 @@ import org.codehaus.plexus.component.annotations.Requirement;
 import org.fusesource.jansi.AnsiConsole;
 
 import java.io.FileOutputStream;
+import java.nio.file.Paths;
 import java.util.Properties;
 import java.util.function.Consumer;
 
 import static java.lang.String.join;
-import static nu.aron.nextbuildnumber.Constants.COMMIT;
-import static nu.aron.nextbuildnumber.Constants.VERSION;
-import static nu.aron.nextbuildnumber.Constants.log;
-import static nu.aron.nextbuildnumber.CurrentWorkingDirectory.getCwd;
+import static nu.aron.next.Constants.ARTIFACT_ID;
+import static nu.aron.next.Constants.COMMIT;
+import static nu.aron.next.Constants.GROUP_ID;
+import static nu.aron.next.Constants.NEXT_COMMIT;
+import static nu.aron.next.Constants.NEXT_VERSION;
+import static nu.aron.next.Constants.VERSION;
+import static nu.aron.next.Constants.log;
+import static nu.aron.next.Constants.logError;
+import static nu.aron.next.CurrentWorkingDirectory.getCwd;
 
 /**
  * Queries the deployment repo for current latest version.
@@ -59,11 +65,11 @@ public class NextBuildNumberLifecycleParticipant extends AbstractMavenLifecycleP
         var pom = session.getRequest().getPom().getAbsoluteFile();
         var model = modelFromFile(pom, modelReader);
         checkModel(model);
-        model.getProperties().put(COMMIT, session.getSystemProperties().get(COMMIT));
+        model.getProperties().put(NEXT_COMMIT, session.getSystemProperties().get(NEXT_COMMIT));
         var version = manuallyBumped(model.getVersion(), getCurrent(session, model));
         log("Latest released version {}", version);
         var nextVersion = newVersion(version, branchName(getCwd(session)), 1);
-        session.getSystemProperties().setProperty(VERSION, nextVersion);
+        session.getSystemProperties().setProperty(NEXT_VERSION, nextVersion);
         log("Next version {}", nextVersion);
         saveValues(nextVersion, session, model);
         findModels(List.of(model), modelReader).forEach(m -> persistVersion(nextVersion, m));
@@ -85,9 +91,16 @@ public class NextBuildNumberLifecycleParticipant extends AbstractMavenLifecycleP
 
     private void saveValues(String nextVersion, MavenSession session, Model model) {
         var p = new Properties();
-        p.put("commit", session.getSystemProperties().get("nextversion.commit"));
-        p.put("version", nextVersion);
+        p.put(COMMIT, session.getSystemProperties().get(NEXT_COMMIT));
+        p.put(VERSION, nextVersion);
+        p.put(ARTIFACT_ID, model.getArtifactId());
+        p.put(GROUP_ID, groupIdFromModel(model));
         p.put("gav", join(":", groupIdFromModel(model), model.getArtifactId(), nextVersion));
+        var target = Paths.get("target").toFile();
+        if (!target.mkdir() && !target.exists()) {
+            logError("Unable to create target directory!");
+            throw new PluginException(new IllegalStateException("Unable to create target directory"));
+        }
         Try.run(() -> p.store(new FileOutputStream("target/nextversion.properties"), null))
                 .onFailure(e -> log("Failed to write ", e.getMessage()));
     }
